+++ src/components/tile/tile-footer/tile-footer.d.ts
@@ -3,7 +3,7 @@ import { PaddingProps } from "styled-system";
 
 export interface TileFooterProps extends PaddingProps {
   /** set which background color variant should be used */
-  variant?: "default" | "transparent";
+  variant?: "default" | "black" | "transparent";
 }
 
 declare function TileFooter(