+++ src/components/tile/tile-footer/tile-footer.style.js
@@ -2,13 +2,22 @@ import styled, { css } from "styled-components";
 import { padding } from "styled-system";
 import { baseTheme } from "../../../style/themes";
 
+const getBackgroundColor = (variant) => {
+  switch (variant) {
+    case "transparent":
+      return "transparent";
+    case "black":
+      return "var(--colorsUtilityYin100)";
+    default:
+      return "var(--colorsUtilityMajor100)";
+  }
+};
+
 const StyledTileFooter = styled.div`
   ${padding}
 
   ${({ variant }) => css`
-    background: ${variant === "transparent"
-      ? "transparent"
-      : "var(--colorsUtilityMajor025)"};
+    background: ${getBackgroundColor(variant)};
     border-top: 1px solid var(--colorsUtilityMajor100);
   `}
 `;